chore: refactor earn team stuff out of token list
